Feature additions
Added repeatLastFunction() so that way a function can be repeated quickly.  For example undo could be a certain slice, but then by putting this function on the middle, one can continuously send undo multiple times by just pressing the pie hotkey.   The first argument is the timeOut in seconds how long the previous function is remembered.