64-bit and Clickable functions
The pie menus were verified to work with the AutoHotKey Unicode 64-bit.  The ability to click functions as added as an activation criteria alongside the criteria of focusing a selection and deactivating the pie menu while hovered over.   functionLaunchMode setting was added to the global settings in settings.json and the annotated settings file was updated.  This allows three modes of function activation that can be viewed in the settings-annotated.txt